Migrated functions out of 2_Temporal_Data file into df_manip_plotting.py Setup relative import within 2_Temporal_Data.py (see notes) Deleted old background Developed import code for files (see test_map.py as ex) to avoid relative imports
